requirements.txt
asgiref==3.8.1
attrs==23.2.0
colorama==0.4.6
coverage==7.5.3
Django==5.0.6
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
